Adds in the test for all impl and small fixes on Custom Repo&Custom Service Impl

+++ src/main/java/com/example/guesthousebookingsystem/services/CustomerServiceImpl.java
@@ -40,10 +40,14 @@ public class CustomerServiceImpl implements CustomerService {
         customerRepository.save(customer);
 
     }
-
     @Override
     public void delete(Long id) {
+        if (customerRepository.existsByBookingsCustomerId(id)) {
+            throw new RuntimeException("Kan inte ta bort kund med aktiva bokningar!");
+        }
         customerRepository.deleteById(id);
-
     }
+
+
 }
+
